refactor gcp.py
